Fix: nav bar (bugs and "demo mode" support)
Bugs
* Logo no longer collapses to 0px x 0px
* iPhone viewport no longer persists for desktop-mode stories

Demo Mode support
"Demo mode" will give anonymous users read-only access so they can nose
around the program before creating an account. This means there are now
three states for a user - logged in, logged out and demo mode.

Since "isLoggedIn" is no longer the deciding factor, the header
component now accepts a string, "loggedInAs", instead. The idea is that
this string will be the username of logged in users, or "DEMO" for demo
moders, but undefined for logged out.